+++ website/database/problem_of_day.php
@@ -3,6 +3,7 @@
 require_once "../modules/handler.php";
 
 // IMPORTS
+add("arguments");
 add("environment");
 add("database");
 add("test");
@@ -11,27 +12,22 @@ add("test");
 header('Content-Type: application/json');
 header('Access-Control-Allow-Origin: *');
 
-// CONFIGURE ENVIRONMENT
-$ENV = load(["DB_HOST", "DB_USER", "DB_PASSWORD", "DB_NAME"]);
-
 // CONNECT TO THE DATABASE
-$database = connect($ENV["DB_HOST"], $ENV["DB_USER"], $ENV["DB_PASSWORD"], $ENV["DB_NAME"]);
+$database = database_connect($ENV["DB_HOST"], $ENV["DB_USER"], $ENV["DB_PASSWORD"], $ENV["DB_NAME"]);
 
 // CHECK ARGUMENT
-test('/^[a-z]{2}$/', $_GET["lang"], "lang");
+test('/^[a-z]{2}$/', $ARG["LANG"], "LANG");
 
 // GET TOTAL NUMBER OF PROBLEMS
-$total = intval((request('SELECT COUNT(*) as total FROM problems', null, false, $database)->fetch_assoc())['total']);
+$total = intval((database_request('SELECT COUNT(*) AS total FROM problems', $database)->fetch_assoc())['total']);
 
 // DETERMINE THE PROBLEM OF THE DAY
 $number = crc32(date('Y-m-d')) % $total;
 
-// GET PROBLEM
-$result = request("SELECT * FROM problems LIMIT $number, 1", null, false, $database);
+// GET PROBLEM AND PRINT
+$result = database_request('SELECT * FROM problems LIMIT ' . $number . ', 1', $database);
 
-// DUMP AND PROCESS DATA
-$data = json_decode(($result->fetch_all(MYSQLI_ASSOC))[0]['data_' . $_GET['lang']], true);
-echo json_encode($data);
+echo json_encode(json_decode(($result->fetch_all(MYSQLI_ASSOC))[0]['data_' . $ARG['LANG']], true));
 
 // CLOSE CONNECTION
 $database->close();