+++ website/modules/arguments.php
@@ -1,0 +1,7 @@
+<?php 
+// GET ALL ARGUMENTS
+global $ARG;
+foreach ($_GET as $key => $value) {
+    $ARG[strtoupper($key)] = $value;
+}
+?>