+++ website/modules/database.php
@@ -1,25 +1,21 @@
 <?php 
 // FUNCTION TO CONNECT TO DATABASE
-function connect($host, $user, $password, $name) {
+function database_connect($host, $user, $password, $name) {
     // CONNECT TO DATABASE
     $db = new mysqli($host, $user, $password, $name);
     // CHECK CONNECTION
     if ($db->connect_error) {
-        throw new databaseConnectException("");
+        throw new databaseConnectException();
     }
     // RETURN DATABASE
     return $db;
 }
 
 // FUNCTION TO REQUEST DATA
-function request($query, $parameter, $wildcard, $db) {
+function database_request($query, $db) {
+    // PREPARE QUERY
     $request = $db->prepare($query);
-    if ($wildcard) {
-        $parameter = "%" . $parameter . "%";
-    }
-    if ($parameter) {
-        $request->bind_param("s", $parameter);
-    }
+    // EXECUTE, GET AND RETURN RESULT
     $request->execute();
     $result = $request->get_result();
     $request->close();