+++ website/modules/handler.php
@@ -1,42 +1,29 @@
 <?php 
-// INPUT VALIDATION ERROR
-class regexException extends Exception {
-    public function errorMessage() {
-        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
-    }
-}
+// REGEX CHECK ERROR
+class regexException extends Exception {}
 // DATABASE CONNECTION ERROR
-class databaseConnectException extends Exception {
-    public function errorMessage() {
-        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
-    }
-}
-// ENVIRONMENT NOT FOUND
-class environmentNotFoundException extends Exception {
-    public function errorMessage() {
-        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
-    }
-}
-// ENVIRONMENT VARIABLE NOT FOUND
-class environmentVariableNotFoundException extends Exception {
-    public function errorMessage() {
-        return "Error ({$this->getFile()}:{$this->getLine()}) -> {$this->getMessage()}";
-    }
-}
+class databaseConnectException extends Exception {}
+// ENVIRONMENT NOT FOUND ERROR
+class environmentNotFoundException extends Exception {}
+// ENVIRONMENT VARIABLE NOT FOUND ERROR
+class environmentVariableNotFoundException extends Exception {}
+// MODULE NOT FOUND ERROR
+class moduleNotFoundException extends Exception {}
 
 // EXCEPTION HANDLER
 function exceptionHandler(Throwable $exception) {
-    echo $exception->getMessage();
+    echo "Error ({$exception->getFile()}:{$exception->getLine()}:{" . get_class($exception) . "}) -> {$exception->getMessage()}";
+    exit;
 }
 set_exception_handler("exceptionHandler");
 
 // ADD MODULE FUNCTION
 function add($module) {
-    try {
-        $path = __DIR__ . "/../modules/$module.php";
+    $path = __DIR__ . "/../modules/$module.php";
+    if (!file_exists($path)) {
+        throw new moduleNotFoundException($module);
+    } else {
         require_once $path;
-    } catch (Exception $error) {
-        echo "Failed to import $module.";
     }
 }
 ?>