+++ website/modules/settings.php
@@ -2,5 +2,5 @@
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <link rel="icon" href="../public/logo.svg" type="svg">
-<meta name="author" content="Jorge Briñas, Beltrán Chicharro, Diego Silva, Paola Leonis y Alejandro Vaz.">
+<meta name="author" content="Jorge Briñas y Alejandro Vaz.">
 <script src="../script/general.js"></script>